Improved (hopefully) the robot navigation
1.  Reduced the refractory period between sonar reads from 10 ms to 3
ms.  As it stood, the robot might travel as much as 1 inch in the time
it took to check all sonars in the move() and followWall() methods.
This should make navigation more precise.
2. In followWall(), added a parameter to allow the robot to follow the
wall with just one sensor.  In this case, we can't correct for rotation,
but at least we can keep the robot from drifting into one of the walls
as it moves across the intersection.
3. Added a MoveBy condition that allows the robot to move a precise
distance forward.
4. Collected UV data from all sensors during a rotate and had rotate()
return the sensor that showed the max change in UV.  This allows us to
enter a room and do a scan to determine the corner holding the candle
before doing any more movement within the room.
5. Added navigation code to make use of these changes for rooms 1, 2,
and 3.   Didn't have time to do it for room 4 (the little room)